Fix For Dashboard Setting QS Param
